requirements.txt
Pillow>=9.0.0
numpy>=1.21.0
brotli>=1.0.9
pytest>=7.0.0
zstandard>=0.25.0
click>=8.1.0
pdf2image>=1.16.0
python-pptx>=0.6.21
beautifulsoup4>=4.9.0